#do not use them

+++ EMProject/MathHelper.h
@@ -7,7 +7,7 @@ namespace em {
 
 		using System::Math;
 		interface class MathHelper{
-			static double EPSILON = std::numeric_limits<double>::epsilon();
+			static double EPSILON = std::numeric_limits<float>::epsilon();
 
 			static double degreeToRadian(Scalar^ d);
 			static Scalar^ radianToDegree(double r);
